requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
diffusers==0.32.0.dev0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
Jinja2==3.1.3
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==10.2.0
pip==24.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.25.0
scipy==1.14.1
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==65.5.0
six==1.17.0
smmap==5.0.1
sympy==1.13.1
tifffile==2024.12.12
tokenizers==0.21.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.1
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
absl-py==2.1.0
grpcio==1.69.0
markdown==3.7
tensorboard==2.18.0
tensorboard-data-server==0.7.2
werkzeug==3.1.3
